+++ Runtime/Scripts/Enums/SQLiteAndroidBasePath.cs
@@ -1,0 +1,7 @@
+namespace ParkMinPackages.SQLiteToolkit
+{
+	public enum SQLiteAndroidBasePath
+	{
+		PersistentDataPath
+	}
+}