+++ Runtime/Scripts/Enums/SQLiteEditorBasePath.cs
@@ -1,0 +1,8 @@
+namespace ParkMinPackages.SQLiteToolkit
+{
+	public enum SQLiteEditorBasePath
+	{
+		PersistentDataPath,
+		Assets
+	}
+}