+++ Runtime/Scripts/Enums/SQLiteStandaloneBasePath.cs
@@ -1,0 +1,9 @@
+namespace ParkMinPackages.SQLiteToolkit
+{
+	public enum SQLiteStandaloneBasePath
+	{
+		PersistentDataPath,
+		ExecutableDirectory,
+		DataDirectory
+	}
+}