+++ Runtime/Scripts/Enums/SQLiteWebGLBasePath.cs
@@ -1,0 +1,7 @@
+namespace ParkMinPackages.SQLiteToolkit
+{
+	public enum SQLiteWebGLBasePath
+	{
+		PersistentDataPath
+	}
+}