+++ Runtime/Scripts/Interfaces/ISQLiteRecord.cs
@@ -1,0 +1,7 @@
+namespace ParkMinPackages.SQLiteToolkit
+{
+	public interface ISQLiteRecord<TKey>
+	{
+		TKey Id { get; }
+	}
+}